+++ public/mod/remove_user.php
@@ -1,11 +1,18 @@
 <?php
 require '../../res/config.php';
 
-$username = $_POST['username'];
+// Check for valid login before doing anything.
+session_start();
+if ($_SESSION['permission'] == 'admin') {
+    $username = $_POST['username'];
 
-$query = $db->prepare("delete from users where username = :username");
-$query->bindValue(':username', $username);
-$query->execute();
+    $query = $db->prepare("delete from users where username = :username");
+    $query->bindValue(':username', $username);
+    $query->execute();
 
-header("Location: /mod.php");
+    header("Location: /mod.php");
+}
+else {
+    echo "Invalid login.";
+}
 ?>
